Refactoring and cleanup as requested per PR. Changed some tabs to spaces.
