+++ MekHQ/src/mekhq/campaign/AwardSet.java
@@ -1,9 +1,16 @@
 package mekhq.campaign;
 
+import mekhq.campaign.personnel.Award;
+
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import java.util.Set;
 
+/**
+ * This class represents an award set.
+ * @author Miguel Azevedo
+ *
+ */
 @XmlRootElement(name="awards")
 public class AwardSet{
